+++ backend/src/app.js
@@ -2,8 +2,7 @@ import express from "express";
 import createError from "http-errors";
 import logger from "morgan";
 
-import indexRouter from "./routes/index.js";
-import usersRouter from "./routes/users.js";
+import v1Router from "./routes/v1/index.js";
 
 const app = express();
 
@@ -11,13 +10,15 @@ app.use(logger("dev"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-app.use("/", indexRouter);
-app.use("/users", usersRouter);
+app.use("/api/v1", v1Router);
+
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
   next(createError(404));
 });
+
+app.get("/", (req, res) => res.send("API GuessThePlace"));
 
 // error handler
 app.use(function (err, req, res, next) {
