+++ backend/src/controllers/README.md
@@ -1,0 +1,7 @@
+# Le dossier controllers
+
+Ici sont définies les fonctions qui gèrent la logique des routes.  
+Les fichiers de routes les appellent pour exécuter les actions.  
+Elles servent aussi à valider les données et gérer les erreurs.
+
+On peut créer un fichier
