+++ backend/src/controllers/postsController.js
@@ -1,0 +1,1 @@
+// ici sera implémenté le controller postsController.js
