+++ backend/src/middlewares/README.md
@@ -1,0 +1,5 @@
+# Le dossier middlewares
+
+Ce dossier contient les fonctions exécutées entre la requête et la réponse.  
+Elles servent par exemple à vérifier l’authentification, valider des données ou logger des informations.  
+Elles sont utilisées dans les routes pour ajouter des traitements avant d’appeler les controllers.
