+++ backend/src/models/README.md
@@ -1,0 +1,5 @@
+# Le dossier models
+
+Ce dossier contient les modèles de données.  
+Ils définissent la structure des documents en base et servent d’interface avec la base de données.  
+Les controllers les utilisent pour lire, créer, modifier ou supprimer des données.
