+++ backend/src/routes/v1/README.md
@@ -1,0 +1,6 @@
+# Le dossier routes
+
+Ce dossier contient la définition des endpoints de l’API.  
+Chaque fichier de route correspond à une ressource (ex. posts, users).  
+Les routes reçoivent les requêtes et appellent les controllers associés.  
+Elles permettent aussi d’organiser facilement le versionnement de l’API.
