--- package.json
@@ -1,21 +1,0 @@
-{
-  "name": "guess-the-place",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "start": "node ./bin/start",
-    "dev": "DEBUG=guess-the-place:* nodemon ./bin/start"
-  },
-  "dependencies": {
-    "debug": "^4.4.1",
-    "dotenv": "^17.2.3",
-    "express": "^5.1.0",
-    "http-errors": "^2.0.0",
-    "mongoose": "^8.19.2",
-    "morgan": "^1.10.1"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.10"
-  }
-}
