--- routes/index.js
@@ -1,9 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.get("/", function (req, res, next) {
-  res.send("Ignition!");
-});
-
-export default router;
